.gitconfig core.whitespace = tab-in-indent
core.whitespace
Git comes preset to detect and fix some whitespace issues. It can look
for six primary whitespace issues – three are enabled by default and can
be turned off, and three are disabled by default but can be activated.

The three that are turned on by default are blank-at-eol, which looks
for spaces at the end of a line; blank-at-eof, which notices blank lines
at the end of a file; and space-before-tab, which looks for spaces
before tabs at the beginning of a line.

The three that are disabled by default but can be turned on are
indent-with-non-tab, which looks for lines that begin with spaces
instead of tabs (and is controlled by the tabwidth option);
tab-in-indent, which watches for tabs in the indentation portion of a
line; and cr-at-eol, which tells Git that carriage returns at the end of
lines are OK.